# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/booking-drafts": {
      "get": {
        "description": "Contiene datos personales e identificadores secretos. No utilizar desde el frontend público.",
        "operationId": "BookingsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "headers": {
              "Cache-Control": {
                "description": "El borrador no debe cachearse.",
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDraftResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Listar borradores (gestión)",
        "tags": [
          "Borradores"
        ]
      },
      "post": {
        "description": "Crear con sportId activo. Los demás campos son opcionales al comenzar, pero si se incluyen deben ser válidos. Zona antes de sede; sede y fecha antes de turno. Guardar el UUID devuelto como secreto de acceso. No confirma reservas.",
        "operationId": "BookingsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "headers": {
              "Cache-Control": {
                "description": "El borrador no debe cachearse.",
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDraftResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "409": {
            "description": "Duplicado o registro con dependencias que impiden la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Duplicado o registro con dependencias que impiden la operación."
                }
              }
            }
          },
          "413": {
            "description": "El JSON supera 16 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 413,
                  "message": "El JSON supera 16 KiB."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Iniciar una solicitud",
        "tags": [
          "Borradores"
        ]
      }
    },
    "/api/booking-drafts/{id}": {
      "get": {
        "description": "El UUID actúa como secreto de acceso, sin login. La lectura no revalida la disponibilidad.",
        "operationId": "BookingsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID secreto del borrador. Usar únicamente uno propio; no publicarlo.",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "headers": {
              "Cache-Control": {
                "description": "El borrador no debe cachearse.",
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDraftResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Recuperar una solicitud propia",
        "tags": [
          "Borradores"
        ]
      },
      "patch": {
        "description": "Actualización parcial; omitir campos sin cambios, no enviar null. Cambiar deporte o zona elimina la sede previa si queda incompatible, salvo que se envíe una sede nueva válida. Cambiar deporte, zona, sede o fecha elimina el turno anterior salvo que se envíe uno nuevo válido. El contacto permanece. El cliente debe reemplazar su borrador con la respuesta completa, sin conservar campos eliminados.",
        "operationId": "BookingsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID secreto del borrador. Usar únicamente uno propio; no publicarlo.",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "headers": {
              "Cache-Control": {
                "description": "El borrador no debe cachearse.",
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDraftResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "409": {
            "description": "Duplicado o registro con dependencias que impiden la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Duplicado o registro con dependencias que impiden la operación."
                }
              }
            }
          },
          "413": {
            "description": "El JSON supera 16 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 413,
                  "message": "El JSON supera 16 KiB."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Guardar un paso de la solicitud",
        "tags": [
          "Borradores"
        ]
      },
      "delete": {
        "description": "Elimina solo el borrador. No es una cancelación de una reserva confirmada por WhatsApp.",
        "operationId": "BookingsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID secreto del borrador. Usar únicamente uno propio; no publicarlo.",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado. Sin cuerpo de respuesta."
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "409": {
            "description": "Duplicado o registro con dependencias que impiden la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Duplicado o registro con dependencias que impiden la operación."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Descartar una solicitud propia",
        "tags": [
          "Borradores"
        ]
      }
    },
    "/api/booking-drafts/{id}/confirm": {
      "post": {
        "description": "Solo administración. Requiere solicitud completa y turno futuro disponible. Confirmación atómica e idempotente; otra solicitud sobre el mismo turno recibe 409. Una solicitud confirmada no puede editarse ni eliminarse.",
        "operationId": "BookingsController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID secreto del borrador. Usar únicamente uno propio; no publicarlo.",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "headers": {
              "Cache-Control": {
                "description": "El borrador no debe cachearse.",
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDraftResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "409": {
            "description": "Duplicado o registro con dependencias que impiden la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Duplicado o registro con dependencias que impiden la operación."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Confirmar y reservar un turno",
        "tags": [
          "Borradores"
        ]
      }
    },
...
  },
  "info": {
    "title": "API del turnero de polideportivos",
    "description": "Flujo: deporte → contacto → zona y sede → detalle → fecha y turno → vista previa de WhatsApp. Los catálogos empiezan vacíos. Cargarlos con clave de gestión en el orden deportes, zonas, sedes, relaciones sede/deporte y turnos. Los UUID de los ejemplos son ficticios. Generar WhatsApp no envía mensajes, no bloquea disponibilidad y no confirma reservas. Los borradores se acceden mediante UUID secreto, sin login. JSON de hasta 16 KiB; 100 solicitudes/minuto/IP.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "ManagementKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Clave de gestión solo para operaciones protegidas. No se almacena en la documentación."
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "Código de estado HTTP."
          },
          "message": {
            "description": "Mensaje de negocio o lista de errores de validación.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "El turno no está disponible"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Nombre del error; puede omitirse, por ejemplo en el límite de solicitudes."
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "BookingDraftResponseDto": {
        "type": "object",
        "properties": {
          "sportId": {
            "type": "string",
            "format": "uuid",
            "description": "Deporte activo elegido; único campo obligatorio al crear",
            "example": "00000000-0000-4000-8000-000000000001"
          },
          "renterFirstName": {
            "type": "string",
            "description": "Nombre del solicitante. Texto plano; se eliminan espacios extremos.",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]*$",
            "example": "Ana"
          },
          "renterLastName": {
            "type": "string",
            "description": "Apellido del solicitante. Texto plano; se eliminan espacios extremos.",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[^<>\\u0000-\\u001f\\u007f]*$",
            "example": "Ejemplo"
          },
          "renterPhone": {
            "type": "string",
            "description": "Teléfono internacional con + y 8–15 dígitos, sin espacios.",
            "pattern": "^\\+[1-9]\\d{7,14}$",
            "example": "+5491100000000"
          },
          "zoneId": {
            "type": "string",
            "format": "uuid",
            "description": "Zona seleccionada; requerida antes de elegir sede",
            "example": "00000000-0000-4000-8000-000000000001"
          },
          "venueId": {
            "type": "string",
            "format": "uuid",
            "description": "Sede habilitada compatible con zona y deporte",
            "example": "00000000-0000-4000-8000-000000000001"
          },
          "slotId": {
            "type": "string",
            "format": "uuid",
            "description": "Turno futuro y disponible de la sede, deporte y fecha seleccionados",
            "example": "00000000-0000-4000-8000-000000000001"
          },
          "date": {
            "type": "string",
            "format": "date",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "description": "Día local en America/Argentina/Buenos_Aires.",
            "example": "2099-01-01"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING_CONFIRMATION",
              "CONFIRMED"
            ]
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "confirmedAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID secreto de acceso al borrador. Quien lo conoce puede leerlo, modificarlo y eliminarlo. No publicarlo ni registrarlo.",
            "example": "00000000-0000-4000-8000-000000000001"
          }
        },
        "required": [
          "sportId",
          "status",
          "id"
        ]
      },
...
    }
  }
}